backend/requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
cryptography>=41.0.0
requests>=2.31.0
django-cors-headers>=4.3.0
drf-spectacular>=0.27.0
